refactor: centralize value helpers
