+++ src/backend_ast/interpreter.c
@@ -2518,325 +2518,6 @@ void executeWithScope(AST *node, bool is_global_scope)  {
     }
 }
 
-/*
-int computeFlatOffset(Value *array, int *indices) {
-    int offset = 0;
-    int multiplier = 1;
-
-    for (int i = array->dimensions - 1; i >= 0; i--) {
-        int idx = indices[i];
-        int lb = array->lower_bounds[i];
-        int ub = array->upper_bounds[i];
-
-        if (idx < lb || idx > ub) {
-            fprintf(stderr, "Runtime error: Index %d out of bounds [%d..%d] in dimension %d.\n",
-                    idx, lb, ub, i + 1);
-            EXIT_FAILURE_HANDLER();
-        }
-
-        offset += (idx - lb) * multiplier;
-        multiplier *= (ub - lb + 1);
-    }
-
-    return offset;
-}
- */
-
-Value makeCopyOfValue(const Value *src) {
-    Value v;
-    v = *src;  // shallow copy to start
-
-    switch (src->type) {
-        case TYPE_STRING:
-            if (src->max_length > 0) {
-                v.s_val = malloc(src->max_length + 1);
-                if (!v.s_val) {
-                    fprintf(stderr, "Memory allocation failed in makeCopyOfValue (string)\n");
-                    EXIT_FAILURE_HANDLER();
-                }
-                if (src->s_val) {
-                    strncpy(v.s_val, src->s_val, src->max_length);
-                    v.s_val[src->max_length] = '\0';
-                } else {
-                    v.s_val[0] = '\0';
-                }
-                v.max_length = src->max_length;
-            } else if (src->s_val) {
-                size_t len = strlen(src->s_val);
-                v.s_val = malloc(len + 1);
-                if (!v.s_val) {
-                    fprintf(stderr, "Memory allocation failed in makeCopyOfValue (string)\n");
-                    EXIT_FAILURE_HANDLER();
-                }
-                strcpy(v.s_val, src->s_val);
-            } else {
-                v.s_val = NULL;
-            }
-            break;
-        case TYPE_ENUM:
-            // Perform deep copy for the enum name string
-            v.enum_val.enum_name = src->enum_val.enum_name ? strdup(src->enum_val.enum_name) : NULL;
-            if (src->enum_val.enum_name && !v.enum_val.enum_name) {
-                 fprintf(stderr, "Memory allocation failed in makeCopyOfValue (enum name strdup)\n");
-                 EXIT_FAILURE_HANDLER();
-            }
-            break;
-        case TYPE_RECORD: {
-            FieldValue *head = NULL, *tail = NULL;
-            for (FieldValue *cur = src->record_val; cur; cur = cur->next) {
-                FieldValue *copy = malloc(sizeof(FieldValue));
-                copy->name = strdup(cur->name);
-                copy->value = makeCopyOfValue(&cur->value);
-                copy->next = NULL;
-                if (tail)
-                    tail->next = copy;
-                else
-                    head = copy;
-                tail = copy;
-            }
-            v.record_val = head;
-            break;
-        }
-        case TYPE_ARRAY: {
-            int total = 1;
-            // <<< ADD/VERIFY THIS >>>
-            v.dimensions = src->dimensions;
-            // <<< END ADD/VERIFY >>>
-
-            // Check if dimensions > 0 before mallocing bounds/data
-            if (v.dimensions > 0 && src->lower_bounds && src->upper_bounds) {
-                v.lower_bounds = malloc(sizeof(int) * src->dimensions);
-                v.upper_bounds = malloc(sizeof(int) * src->dimensions);
-                // Add null checks for malloc results
-                if (!v.lower_bounds || !v.upper_bounds) { /* Handle error */ }
-
-                for (int i = 0; i < src->dimensions; i++) {
-                    v.lower_bounds[i] = src->lower_bounds[i];
-                    v.upper_bounds[i] = src->upper_bounds[i];
-                    // Check for valid bounds before calculating total_size
-                    int size_i = (v.upper_bounds[i] - v.lower_bounds[i] + 1);
-                    if (size_i <= 0) { total = 0; break; } // Invalid bounds -> empty array effectively
-                    if (__builtin_mul_overflow(total, size_i, &total)) { total = -1; break; } // Check overflow
-                }
-            } else {
-                // If dimensions is 0 or bounds are missing, ensure size is 0
-                total = 0;
-                v.lower_bounds = NULL;
-                v.upper_bounds = NULL;
-            }
-
-            // Allocate and copy array data only if total size is valid
-            v.array_val = NULL; // Initialize to NULL
-            if (total > 0 && src->array_val) {
-                 v.array_val = malloc(sizeof(Value) * total);
-                 if (!v.array_val) { /* Handle error */ }
-                 for (int i = 0; i < total; i++) {
-                     v.array_val[i] = makeCopyOfValue(&src->array_val[i]);
-                 }
-            } else if (total < 0) {
-                 fprintf(stderr, "Error: Array size overflow during copy.\n");
-                 // Handle error - maybe set dimensions to 0?
-                 v.dimensions = 0;
-                 free(v.lower_bounds); v.lower_bounds = NULL;
-                 free(v.upper_bounds); v.upper_bounds = NULL;
-            } else {
-                 // total is 0, ensure array_val is NULL (already done)
-            }
-
-            // Copy element type definition link
-            v.element_type_def = src->element_type_def; // Assuming shallow copy is ok for AST node link
-            v.element_type = src->element_type; // Copy element type enum
-
-            break;
-        } // End case TYPE_ARRAY
-        case TYPE_CHAR:
-            // Already handled by shallow copy
-            break;
-        case TYPE_SET:
-            // Deep copy the set elements
-            v.set_val.set_values = NULL; // Initialize destination pointer
-            v.set_val.set_size = 0;      // Initialize destination size
-            // v.max_length = 0; // Reset capacity tracking if you use it
-
-            if (src->set_val.set_size > 0 && src->set_val.set_values != NULL) {
-                size_t array_size_bytes = sizeof(long long) * src->set_val.set_size;
-                v.set_val.set_values = malloc(array_size_bytes);
-                if (!v.set_val.set_values) {
-                    freeValue(&v);
-                    fprintf(stderr,
-                            "Memory allocation failed in makeCopyOfValue (set)\n");
-                    EXIT_FAILURE_HANDLER();
-                }
-                if (!v.set_val.set_values) {
-                    fprintf(stderr, "Memory allocation failed in makeCopyOfValue (set values)\n");
-                    // Potentially free other copied parts before exiting if necessary
-                    EXIT_FAILURE_HANDLER();
-                }
-                memcpy(v.set_val.set_values, src->set_val.set_values, array_size_bytes);
-                v.set_val.set_size = src->set_val.set_size;
-                 // Optionally copy capacity if tracked: v.max_length = src->max_length;
-            }
-            // If source size is 0 or values pointer is NULL, destination remains empty (NULL/0)
-            break;
-        default:
-            // ints, bools, reals etc. are already safely copied by value
-            break;
-    }
-
-    return v;
-}
-Value* resolveLValueToPtr(AST* lvalueNode) {
-    if (!lvalueNode) {
-        fprintf(stderr, "Runtime error: Cannot resolve NULL lvalue node.\n");
-        EXIT_FAILURE_HANDLER();
-    }
-
-    switch (lvalueNode->type) {
-        case AST_VARIABLE: {
-            Symbol* sym = lookupSymbol(lvalueNode->token->value); // Handles not found error
-            if (sym->is_const) {
-                 fprintf(stderr, "Runtime error: Cannot modify constant symbol '%s'.\n", sym->name);
-                 EXIT_FAILURE_HANDLER();
-            }
-            if (!sym->value) {
-                 fprintf(stderr, "Runtime error: Symbol '%s' has NULL value pointer.\n", sym->name);
-                 EXIT_FAILURE_HANDLER();
-            }
-            return sym->value; // Return pointer to the symbol's value storage
-        }
-
-        case AST_ARRAY_ACCESS: {
-            // Recursively resolve the base of the array access (e.g., 'myArray' or 'myRecord.myArray')
-            Value* baseValuePtr = resolveLValueToPtr(lvalueNode->left);
-            if (!baseValuePtr) { /* Error handled in recursive call */ EXIT_FAILURE_HANDLER(); }
-
-            // Check if the resolved base is actually an array or string
-            if (baseValuePtr->type == TYPE_ARRAY) {
-                // --- Array Element Access ---
-                if (!baseValuePtr->array_val) { /* Error: Array not initialized */ }
-#ifdef DEBUG
-fprintf(stderr, "[DEBUG RESOLVE_LVAL] Array Access Check:\n");
-fprintf(stderr, "  lvalueNode->child_count = %d\n", lvalueNode->child_count);
-fprintf(stderr, "  baseValuePtr->dimensions = %d\n", baseValuePtr->dimensions);
-fflush(stderr); // Ensure output is seen before potential crash
-#endif
-                if (lvalueNode->child_count != baseValuePtr->dimensions) { /* Error: Index count mismatch */ }
-
-                // Calculate indices
-                int* indices = malloc(sizeof(int) * baseValuePtr->dimensions);
-                if (!indices) { /* Mem Error */ }
-                for (int i = 0; i < lvalueNode->child_count; i++) {
-                    assert(i < baseValuePtr->dimensions);
-                    Value idxVal = eval(lvalueNode->children[i]); // Eval index expressions
-                    if (idxVal.type != TYPE_INTEGER) { /* Index Type Error */ free(indices); freeValue(&idxVal); EXIT_FAILURE_HANDLER(); }
-                    indices[i] = (int)idxVal.i_val;
-                    freeValue(&idxVal); // Free temp index value
-                }
-
-                // Calculate offset and check bounds
-                int offset = computeFlatOffset(baseValuePtr, indices);
-                int total_size = 1;
-                for (int i = 0; i < baseValuePtr->dimensions; i++) { total_size *= (baseValuePtr->upper_bounds[i] - baseValuePtr->lower_bounds[i] + 1); }
-                 if (offset < 0 || offset >= total_size) { /* Bounds Error */ free(indices); EXIT_FAILURE_HANDLER(); }
-
-                free(indices); // Free indices array after calculating offset
-
-                // Return pointer to the element within the array's value storage
-                return &(baseValuePtr->array_val[offset]);
-
-            } else if (baseValuePtr->type == TYPE_STRING) {
-                // --- String Character Access ---
-                // For L-Value resolution, we return a pointer to the *base string's Value* struct.
-                // The caller (AST_ASSIGN) MUST detect this specific case and handle modification.
-                #ifdef DEBUG
-                const char* base_name = lvalueNode->left->token ? lvalueNode->left->token->value : "?"; // Get base var name for debug
-                fprintf(stderr, "[DEBUG RESOLVE_LVAL] String index access LValue ('%s[?]'). Returning pointer to base string Value* %p\n",
-                        base_name, (void*)baseValuePtr);
-                #endif
-                // NOTE: We don't evaluate or check the index here. The caller must do it.
-                return baseValuePtr; // <<< RETURN POINTER TO BASE STRING VALUE >>> // Modified Line
-            } else {
-                fprintf(stderr, "Runtime error: Attempted array/string access on non-array/string type (%s).\n", varTypeToString(baseValuePtr->type));
-                EXIT_FAILURE_HANDLER();
-            }
-            break; // Keep compiler happy
-        }
-
-        case AST_FIELD_ACCESS: {
-            // Recursively resolve the base of the field access (e.g., 'myRecord' or 'myArray[i]')
-            Value* baseValuePtr = resolveLValueToPtr(lvalueNode->left);
-             if (!baseValuePtr) { /* Error handled in recursive call */ EXIT_FAILURE_HANDLER(); }
-
-            // Check if the resolved base is actually a record
-            if (baseValuePtr->type != TYPE_RECORD) {
-                 fprintf(stderr, "Runtime error: Field access on non-record type (%s).\n", varTypeToString(baseValuePtr->type));
-                 EXIT_FAILURE_HANDLER();
-            }
-             if (!baseValuePtr->record_val) {
-                  fprintf(stderr, "Runtime error: Record accessed before initialization or after being freed.\n");
-                  EXIT_FAILURE_HANDLER();
-             }
-
-            // Find the specific field by name
-            const char* targetFieldName = lvalueNode->token ? lvalueNode->token->value : NULL;
-            if (!targetFieldName) { /* Invalid AST node error */ }
-
-            FieldValue* currentField = baseValuePtr->record_val;
-            while (currentField) {
-                if (currentField->name && strcmp(currentField->name, targetFieldName) == 0) {
-                    // Found the field, return pointer to its Value struct
-                    return &(currentField->value);
-                }
-                currentField = currentField->next;
-            }
-
-            // Field not found
-            fprintf(stderr, "Runtime error: Field '%s' not found in record.\n", targetFieldName);
-            EXIT_FAILURE_HANDLER();
-            break; // Keep compiler happy
-        }
-        case AST_DEREFERENCE: {
-            #ifdef DEBUG
-            fprintf(stderr, "[DEBUG RESOLVE_LVAL] Resolving DEREFERENCE node...\n");
-            #endif
-            // 1. Evaluate the pointer expression itself (node->left)
-            Value pointerValue = eval(lvalueNode->left);
-
-            // 2. Check if it's actually a pointer
-            if (pointerValue.type != TYPE_POINTER) {
-                fprintf(stderr, "Runtime error: Attempting to dereference a non-pointer type (%s) in LValue.\n", varTypeToString(pointerValue.type));
-                freeValue(&pointerValue); // Free temporary value
-                EXIT_FAILURE_HANDLER();
-            }
-
-            // 3. Check if the pointer is nil
-            if (pointerValue.ptr_val == NULL) {
-                fprintf(stderr, "Runtime error: Attempting assignment via a nil pointer.\n");
-                // Don't free pointerValue if it was nil (no heap data)
-                EXIT_FAILURE_HANDLER();
-            }
-
-            // 4. The target for assignment is the Value struct pointed to by ptr_val
-            Value* targetValuePtr = pointerValue.ptr_val;
-
-            // 5. Free the temporary pointerValue struct returned by eval (it just held the address)
-            //    Do NOT free the memory it points to (*targetValuePtr)
-            //    Since pointerValue holds no heap data itself (just the ptr_val address),
-            //    we don't need freeValue here.
-
-            #ifdef DEBUG
-            fprintf(stderr, "[DEBUG RESOLVE_LVAL] Dereference resolved to heap Value* address: %p\n", (void*)targetValuePtr);
-            #endif
-
-            return targetValuePtr; // Return the pointer TO THE HEAP-ALLOCATED VALUE
-        }
-
-        default:
-            fprintf(stderr, "Runtime error: Invalid lvalue node type (%s) for assignment target resolution.\n", astTypeToString(lvalueNode->type));
-            EXIT_FAILURE_HANDLER();
-    }
-    return NULL; // Should not be reached
-}
 
 Value setUnion(Value setA, Value setB) {
     // Basic check: Ensure both are sets
