+++ src/backend_ast/interpreter.h
@@ -5,7 +5,6 @@
 #include "parser.h"
 #include "ast.h"
 
-#define PASCAL_DEFAULT_FLOAT_PRECISION 6
 
 // Execution
 Value eval(AST *node);
@@ -13,16 +12,12 @@ Value executeProcedureCall(AST *node);
 //ExecStatus executeWithScope(AST *node, bool is_global_scope);
 void executeWithScope(AST *node, bool is_global_scope);
 
-Value makeCopyOfValue(const Value *src);
-
 // Memory
 FieldValue *copyRecord(FieldValue *orig);
 FieldValue *createEmptyRecord(AST *recordType);
 
 // Arrays
 int computeFlatOffset(Value *array, int *indices);
-
-Value* resolveLValueToPtr(AST* lvalueNode); // Returns pointer to the Value struct to be modified
 
 // Program flow
 typedef enum {
