+++ src/core/utils.h
@@ -4,6 +4,7 @@
 #include "symbol.h"
 #include "vm.h"
 
+#define PASCAL_DEFAULT_FLOAT_PRECISION 6
 /* =======================
    DEBUG MACROS & GLOBALS
    ======================= */
@@ -154,5 +155,7 @@ int map16BgColorToAnsi(int pscalColorCode);
 // Arrays
 Value makeArrayND(int dimensions, int *lower_bounds, int *upper_bounds, VarType element_type, AST *type_def);
 int computeFlatOffset(Value *array, int *indices);
+Value makeCopyOfValue(const Value *src);
+Value* resolveLValueToPtr(AST* lvalueNode);
 
 #endif // UTILS_H
